# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/assignment-readiness:
    post:
      operationId: generateAssignmentReadiness
      tags: [profile]
      summary: Surgically precise, analytics-based pointers on what to focus on before a graded assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentReadinessInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentReadiness"
components:
  schemas:
...
    ProfileTopicStat:
      type: object
      required: [topicId, topicTitle, weekNumber, attempts, accuracy, avgScore, masteryLabel]
      properties:
        topicId: { type: integer }
        topicTitle: { type: string }
        weekNumber: { type: integer }
        attempts: { type: integer }
        accuracy: { type: number }
        avgScore: { type: number }
        masteryLabel:
          type: string
          enum: [mastered, strong, developing, shaky, weak, untested]
        lastPracticedAt: { type: ["string", "null"], format: date-time }
    AssignmentReadinessInput:
      type: object
      required: [assignmentId]
      properties:
        assignmentId: { type: integer }
    ReadinessPointer:
      type: object
      required: [severity, text]
      properties:
        topicId: { type: ["integer", "null"] }
        topicTitle: { type: ["string", "null"] }
        severity:
          type: string
          enum: [focus, review, solid]
        text: { type: string }
    AssignmentReadiness:
      type: object
      required: [assignmentId, assignmentTitle, readinessScore, readinessLabel, headline, practiceCount, pointers, topics]
      properties:
        assignmentId: { type: integer }
        assignmentTitle: { type: string }
        readinessScore: { type: number }
        readinessLabel:
          type: string
          enum: [ready, almost, keep_practicing, not_ready]
        headline: { type: string }
        practiceCount: { type: integer }
        pointers:
          type: array
          items:
            $ref: "#/components/schemas/ReadinessPointer"
        topics:
          type: array
          items:
            $ref: "#/components/schemas/ProfileTopicStat"